Fix EDIM to work with current GAP master branch
In the GAP master branch, the way we store the length of plists was changed.
The EDIM kernel module did not use the official API to set the length of a
plist it produced, which means it is broken with the current GAP master
branch.

This PR fixes this, while staying fully compatible with older GAP releases.